Refactor partner show page and partner form
This commit refactors the `Show` component in the `Partner` directory and the `PartnerForm` component. It removes unused variables and improves code readability. It also adds validation to the `PartnerForm` and updates the partner's information accordingly. Additionally, it updates the partner's logo if a new one is provided.

+++ resources/views/Tableactions/permits.blade.php
@@ -1,8 +1,11 @@
 <div class="btn-toolbar" role="toolbar" aria-label="Toolbar with button groups">
 
     <div class="btn-group mx-2" role="group" aria-label="Third group">
-      <button type="button" class="btn btn-light" wire:click="$dispatch('delete-user', { id: {{$row->id}}})">
-        <x-heroicon-o-trash style="width: 28px" />
-    </button>
+      
+      @if (havePermission(auth()->user(),'event-delete'))
+          <button type="button" class="btn btn-light" wire:click="$dispatch('delete-user', { id: {{$row->id}}})">
+            <x-heroicon-o-trash style="width: 28px" />
+        </button>
+     @endif
     </div>
   </div>
